Workaround io_pgetevents() for 32-bit userland running on 64-bit kernels
To cope with the broken compat kernel syscall, which uses a 64-bit pointer
instead of a 32-bit one, and eats the sigset_t size member which makes it
then fail a consistency check and always return -EINVAL. This fixes test
case 22.t. The kernel fix is being prepared as well.